Windows platform change: treat char as unsigned.
In Microsoft Visual C++ char types are signed, which can lead to unexpected behaviours e.g. a character value which contains 0xaa, when compared with the literal value 0xaa, will return false; the character value is interpreted as being negative because it has the top bit set, while the literal value 0xaa is positive.  To avoid this problem advice is added that the command-line switch /J should be used with the compiler to force char types to be treated as unsigned.

A check is added to u_port.c that _CHAR_UNSIGNED is defined for MSVC (/J just defines that compilation flag, which has an effect in limits.h) and the CMakeLists.txt for the runner build is updated to define it.

Then all the nasty casts can go away.

+++ cell/src/u_cell_sec_tls.c
@@ -310,7 +310,7 @@ static int32_t setSequence(const uCellSecTlsContext_t *pContext,
                     // a printable ASCII string
                     errorCode = (int32_t) U_ERROR_COMMON_INVALID_PARAMETER;
                     for (size_t x = 0; (x < size) && good; x++) {
-                        good = isprint((int32_t) (uint8_t) *(pBinary + x)) != 0; // *NOPAD*
+                        good = isprint((int32_t) *(pBinary + x)) != 0; // *NOPAD*
                     }
                     if (good) {
                         errorCode = (int32_t) U_ERROR_COMMON_NO_MEMORY;
